Use INIT_CWD defined by npm to find the root path of the project
The root project may not be a git repository.
By exemple when the root project is built in a Docker container in
several stages to use the Docker cache to cache the installation of
all Node.js dependencies in node_modules.
The first stage is used to copy only needed files (package*.json, npm
scripts, etc...) before to run the command `npm install`.
During the second stage, we copy the whole project inside the
container and run it.

Rather to use the git system command to determine the root path of the
project, we use the INIT_CWD defined by npm.