add eslint-plugin-literal-blacklist and configure it
For more info:
https://github.com/kyaido/eslint-plugin-literal-blacklist

For the moment and to avoid mistakes, the `datadoghq.eu` api host is
blacklisted from string literals. `datadoghq.com` is where you'll find
all your metrics.

Explanation:

Because we are in the EU and every Datadog docs says you should use the
.eu, if you use the EU site, we are tempted to use `api.datadoghq.eu`
instead of `api.datadoghq.com` we use at Leboncoin.